+++ src/main/kotlin/dev/aaronhowser/mods/aaron/AaronExtraCodecs.kt
@@ -2,6 +2,8 @@ package dev.aaronhowser.mods.aaron
 
 import com.mojang.serialization.Codec
 import net.minecraft.Util
+import net.minecraft.network.chat.Component
+import net.minecraft.network.chat.MutableComponent
 import org.joml.Vector2d
 
 object AaronExtraCodecs {
@@ -16,5 +18,12 @@ object AaronExtraCodecs {
 
 	val UINT_CODEC: Codec<UInt> =
 		Codec.INT.xmap(Int::toUInt, UInt::toInt)
+
+	val COMPONENT_CODEC: Codec<MutableComponent> =
+		Codec.STRING
+			.xmap(
+				{ str -> Component.Serializer.fromJson(str) },
+				{ comp -> Component.Serializer.toJson(comp) }
+			)
 
 }